docs(adrs): update ADRs to standardize on key decisions and remove working doc verbiage. refactor(vector_db): update imports for VectorRecord and VectorMatch
