+++ docs/developers/adrs/0005-library-first-retrieval-architecture.md
@@ -2,16 +2,21 @@
 
 **Date:** 2025-10-02  
 **Status:** Accepted  
-**Context:** The legacy query processing stack retained numerous bespoke
-components (federated routing, clustering, ranking, synonym expansion) that
-replicate capabilities now available in mature libraries (LangChain,
-FastEmbed, Qdrant). Maintenance burden and feature delivery velocity have both
-suffered as a result.
+**Drivers:** Legacy bespoke components duplicating mature library capabilities, increased maintenance burden, slower feature delivery  
+**Deciders:** AI Docs Platform Team
+
+## Context
+
+The legacy query processing stack retained numerous bespoke components
+(federated routing, clustering, ranking, synonym expansion) that replicate
+capabilities now available in mature libraries (LangChain, FastEmbed, Qdrant).
+Maintenance burden and feature delivery velocity have both suffered as a
+result.
 
 ## Decision
 
 - Adopt a library-first posture across embeddings, vector storage, retrieval
-  orchestration, and RAG generation.  The core implementation will use:
+  orchestration, and RAG generation. The core implementation will use:
   - `fastembed` for embedding generation (CPU optimised)
   - `langchain_qdrant.QdrantVectorStore` for vector persistence and search
   - LangChain retrievers/rerankers for query expansion, grouping, contextual
@@ -28,11 +33,10 @@ suffered as a result.
 
 - Significant reduction in bespoke code, freeing teams to focus on product
   features and evaluation pipelines.
-- Dependency management becomes critical.  Versions are pinned in
-  `pyproject.toml` and documented in the compatibility matrix.  Renovate
+- Dependency management becomes critical. Versions are pinned in
+  `pyproject.toml` and documented in the compatibility matrix. Renovate
   handles grouped upgrade proposals.
-- Performance regressions must be monitored via regression harnesses (see ADR
-  0006) and canary deployments.  Hot paths may still call the Qdrant SDK
+- Performance regressions must be monitored via regression harnesses (see ADR 0006) and canary deployments. Hot paths may still call the Qdrant SDK
   directly if required.
 
 ## Status
